+++ snuba/web/rpc/v1/resolvers/R_ourlogs/resolver_trace_item_table.py
@@ -6,8 +6,11 @@ from sentry_protos.snuba.v1.request_common_pb2 import TraceItemType
 
 from snuba.state import get_int_config
 from snuba.web.rpc.v1.resolvers import ResolverTraceItemTable
-from snuba.web.rpc.v1.resolvers.R_eap_spans.resolver_trace_item_table import (
-    ResolverTraceItemTableEAPSpans,
+from snuba.web.rpc.v1.resolvers.R_eap_items.resolver_trace_item_table import (
+    ResolverTraceItemTableEAPItems,
+)
+from snuba.web.rpc.v1.resolvers.R_ourlogs.old_resolvers.old_resolver_trace_item_table import (
+    OldResolverTraceItemTableOurlogs,
 )
 
 
@@ -19,8 +22,8 @@ class ResolverTraceItemTableOurlogs(ResolverTraceItemTable):
     def resolve(self, in_msg: TraceItemTableRequest) -> TraceItemTableResponse:
         use_new_logs_resolver = bool(get_int_config("use_new_logs_resolver", default=0))
         if use_new_logs_resolver:
-            res = ResolverTraceItemTableEAPSpans().resolve(in_msg)
+            res = ResolverTraceItemTableEAPItems().resolve(in_msg)
             # option 2 at this point convert the timestamp alias
             return res
         else:
-            raise NotImplementedError("todo add back the old one and use it")
+            return OldResolverTraceItemTableOurlogs().resolve(in_msg)
